Move step 2 out of tests dir lol
